Added WiFi connection check
In TimeManager, if function to get time from NTP server is not able
to get it because WiFi is not connected, software blocks there and
also time is lost. Checking connection before try to update
guarantees that time, even if slighty not corret, is kept updated.
If WiFi is down, a task has been created in order to try the
reconnection