Modified pre-start in order to get information from external flash
Information stored in the external flash are used for the first
connection. In case the stored SSID is not reachable, the latter
is reset and a new one can be selected. There were few problems
during the switch from station to access point mode and viceversa:
update of the avalable networks was prevented after a failed
connection. To avoid this situation, the only way is to reset the
whole board.